Fix counter value not reloading from AppState.
When you open the Counter page, increment, then leave and come back, the counter is reset to 0.  However the AppState keeps the increased value.  This leads to confusing results when you click the WinForms button because it returns AppState value, which may be different than the page.  The fix is to re-load AppState Counter value each time you visit Counter page.